Adjusted import structure
